% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.management.R
\name{benchmark.compare.output}
\alias{benchmark.compare.output}
\title{Compare the output of two benchmarks}
\usage{
benchmark.compare.output(pkg.name, branch.orig, branch.new, output.loc, verbose)
}
\description{
Compare the output of two benchmarks
}

