% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.management.R
\name{benchmark.generate.comparison}
\alias{benchmark.generate.comparison}
\title{Write benchmarks out for a comparison branch}
\usage{
benchmark.generate.comparison(pkg.loc = "./",
  pkg.name = tail(splitstr(normalizePath(pkg.loc), "/")[[1]], 1),
  branch = "master", output.loc = "/tmp/", verbose = TRUE,
  datasets = NULL, fs = NULL)
}
\arguments{
\item{datasets}{is the name of the datasets that should be loaded before benchmarking.}
}
\value{
Nothing yet.
}
\description{
This function goes through the entire namespace of a given package, tests if there are benchmarks in the examples code .  If there is a benchmark, the code for that benchmark, and the reulting output are saved.
}
\author{
Alex
 i <- 20
}

