% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.management.R
\name{benchmark.run.comparison}
\alias{benchmark.run.comparison}
\title{Compare the benchmark results from two versions of the repository}
\usage{
benchmark.run.comparison(pkg.loc = "./",
  pkg.name = tail(strsplit(normalizePath(pkg.loc), "/")[[1]], 1),
  branch.orig = "master", branch.new = "dev", output.loc = "/tmp/",
  verbose = TRUE, datasets = NULL)
}
\description{
Note that this needs to launch separate R
}

