% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.graph.R
\name{call.graph.setup}
\alias{call.graph.setup}
\title{Use the proftools output to create a call graph}
\usage{
call.graph.setup(pd)
}
\arguments{
\item{pd}{The processed output of an Rprof session for a block of R code (see example)}
}
\description{
Using the processed output from an Rprof session, this will create an graph object representing the call graph.  It is focused on a particular namespace and colors nodes according to their documentation/review states
}
\examples{
data(market_95)
f <- "pr <- price.response(DF)"
pd <- prof.routine(f)
g <- call.graph.setup(pd)
}
\author{
Alex
}

