% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.management.R
\name{compare.benchmarks}
\alias{compare.benchmarks}
\title{Benchmark all functions in a package}
\usage{
compare.benchmarks(pkg.loc = "./",
  pkg.name = tail(splitstr(normalizePath(pkg.loc), "/")[[1]], 1),
  branch.orig = "master", branch.new = "dev", skip.same = TRUE,
  verbose = TRUE, datasets = NULL)
}
\value{
Nothing yet.
}
\description{
This function goes through the entire namespace of a given package, tests if there are benchmarks in the examples code .  The md5 hash of the function source code as well as the hashed results of any benchmarks are then stored in a new file
}
\author{
Alex
}

