% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.interaction.R
\name{git.current.branch}
\alias{git.current.branch}
\title{Find out all local branches}
\usage{
git.current.branch(loc = "./")
}
\arguments{
\item{loc}{Where to look for the package}
}
\value{
A string with the name of the current branch
}
\description{
Find out all local branches
}

