% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.management.R
\name{hash.obj}
\alias{hash.obj}
\title{Create the md5 hash of an object}
\usage{
hash.obj(obj)
}
\arguments{
\item{obj}{Any R object that can be saved with the save command.}
}
\value{
The md5 hash of an object
}
\description{
R's tools only allow one to create the md5 hash of a file.  This function takes an R object, creates a file, calculates the hash, deletes the file and returns the result.
}
\examples{
## BENCHMARK
set.seed(1)
r <- rnorm(1e5)
r.md5 <- hash.obj(r)
names(r.md5) <- NULL
r.md5
## END BENCHMARK
}
\author{
Alex
}

