% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.management.R
\name{lookup.benchmarks}
\alias{lookup.benchmarks}
\title{Find the currently stored benchmarks and return them}
\usage{
lookup.benchmarks(str.f, package.loc = "~/pkg/pricemethodresearch")
}
\arguments{
\item{str.f}{The function name as a string}

\item{package.loc}{Where the source code for the package is, and benchmarks}
}
\description{
For a given function, lookup the currently stored benchmarks and return them.
}
\examples{
## BENCHMARK
a <- lookup.benchmarks("lookup.benchmarks")
a
## END BENCHMARK
}

