% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.graph.R
\name{prof.routine}
\alias{prof.routine}
\title{A wrapper for proftools to profile a set of commands}
\usage{
prof.routine(f, rep = 1)
}
\arguments{
\item{f}{a string containing R code to be parsed and run.}

\item{rep}{An integer for the number of times to repear the f evaluation, helps capture fast routines}
}
\value{
a proftools parse of the profile data
}
\description{
This is a helper function to run a set of commands in an Rprof'ed session and return the profile data
}
\author{
Alex
}

