% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.graph.R
\name{reduce.p}
\alias{reduce.p}
\title{Helper function to reduce the proftools call graph to limited namespace}
\usage{
reduce.p(p, e)
}
\arguments{
\item{p}{The p object crunched by proftools}

\item{e}{The environment agains which p should be reduced.}
}
\value{
A new object with the exact same form as p, but with only those functions in e
}
\description{
This is a hack that takes an object from the proftools package and slims it down.  Basically the callgraph code in proftools included all sorts of R internals nonsense and had no option to reduce it.  This code does that, in a pretty aggressive way.  Unclear how this could be useful beyond the call.graph.setup routine.
}
\examples{
data(market_95)
f <- "pr <- price.response(DF)"
pd <- prof.routine(f)
}
\author{
Alex
}

