% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.management.R
\name{run.benchmarks}
\alias{run.benchmarks}
\title{Run the benchmark for a specific function and return relevant hashes}
\usage{
run.benchmarks(str.f, package.loc = "~/pkg/pricemethodresearch",
  verbose = TRUE, pkg.name = "pricemethodresearch")
}
\arguments{
\item{str.f}{The function name as a string}

\item{package.loc}{Where the source code for the package is, and benchmarks}

\item{verbose}{Chatty}
}
\description{
For a given function, compute the hash of the function source
as well as the hash of any associated benchmarks and return
}
\examples{
data(market_2)
## BENCHMARK
a <- run.benchmarks("km")
a
## END BENCHMARK
}

