% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.management.R
\name{test.benchmarks}
\alias{test.benchmarks}
\title{Test a set of functions in memory against their pre-stored benchmarks}
\usage{
test.benchmarks(pkg.name = "pricemethodresearch",
  str = ls(as.environment(paste0("package:", pkg.name))),
  benchmark.loc = "~/pkg/pricemethodresearch/", verbose = TRUE)
}
\arguments{
\item{str}{A vector of length n with the names of the functions you'd like to test}

\item{benchmark.loc}{The path where the package source code lives}

\item{verbose}{Tell me about your day}
}
\value{
A vector of length n with the whether the benchmarks passed (TRUE) or at least one failed (FALSE) for each function
}
\description{
This just tests whether the current working version of a function
Lines up against its benchmark
}
\examples{
data(market_2)
## BENCHMARK
km <- function(DF){return(DF)}
a <- test.benchmarks(str = "km") ## Should fail
a
## END BENCHMARK
}

